# colexec: fix some benchmark issues due to a recent change